requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
coremltools==8.3.0
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.1
keras==2.12.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pyaml==25.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
